Split E2E tests into three files
