+++ packages/playground/website/project.json
@@ -136,6 +136,7 @@
 			"options": {
 				"cypressConfig": "packages/playground/website/cypress.config.ts",
 				"testingType": "e2e",
+				"headless": false,
 				"baseUrl": "https://playground.test/website-server/"
 			}
 		}
